Added half of account logic

+++ src/main/java/by/innowise/task/dao/BaseDao.java
@@ -5,7 +5,7 @@ import by.innowise.task.exception.DaoException;
 import java.util.List;
 
 public interface BaseDao<T> {
-    T findById(int id) throws DaoException;
+    T findById(long id) throws DaoException;
     List<T> findAll() throws DaoException;
     void insert(T t) throws DaoException;
     void update(T t) throws DaoException;
